restructure code and rename package to fmlearn
