+++ sklearn/__init__.py
@@ -91,7 +91,7 @@ else:
                'naive_bayes', 'neighbors', 'neural_network', 'pipeline',
                'preprocessing', 'random_projection', 'semi_supervised',
                'svm', 'tree', 'discriminant_analysis', 'impute', 'compose',
-               'fml',
+               'fmlearn',
                # Non-modules:
                'clone', 'get_config', 'set_config', 'config_context',
                'show_versions']
