--- sklearn/fml/encryption/__init__.py
@@ -1,7 +1,0 @@
-"""
-The :mod:`sklearn.fml.encryption` module implements Federated Meta Learning
-"""
-
-from .fml_hash import FMLHash
-
-__all__ = ['FMLHash']
