+++ sklearn/fmlearn/constants.py
