+++ sklearn/fmlearn/encryption/__init__.py
@@ -1,0 +1,7 @@
+"""
+The :mod:`sklearn.fmlearn.encryption` module implements Federated Meta Learning
+"""
+
+from .fml_hash import FMLHash
+
+__all__ = ['FMLHash']
